Extract batch paxos resources
