+++ timelock-impl/src/main/java/com/palantir/atlasdb/timelock/paxos/UseCaseAwareBatchPaxosResource.java
@@ -22,8 +22,6 @@ import javax.ws.rs.NotFoundException;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 
-import org.immutables.value.Value;
-
 import com.google.common.collect.ImmutableMap;
 
 @Path("/" + PaxosTimeLockConstants.INTERNAL_NAMESPACE)
@@ -59,14 +57,6 @@ public class UseCaseAwareBatchPaxosResource {
             throw new NotFoundException();
         }
         return resourcesForUseCase;
-    }
-
-    @Value.Immutable
-    interface BatchPaxosResources {
-        @Value.Parameter
-        BatchPaxosAcceptorResource batchAcceptor();
-        @Value.Parameter
-        BatchPaxosLearnerResource batchLearner();
     }
 
 }
